extract oauth code out of LoginController.cs and into OauthController.cs, revert some changes made to vite.config.ts
