+++ frontend/src/routes/(authenticated)/authorize/+page.svelte
@@ -12,7 +12,7 @@
 <div class="flex flex-col items-center grow">
   <div class="flex flex-col justify-center grow max-w-lg">
     <div class="flex-grow"></div>
-    <form method="post" action="/api/login/open-id-auth" class="flex flex-col items-center">
+    <form method="post" action="/api/oauth/open-id-auth" class="flex flex-col items-center">
       <div class="mb-6 grid gap-x-3 gap-y-1 items-center" style="grid-template-columns: auto 1fr">
         <div class="row-span-2">
           <Icon icon="i-mdi-approval" color="text-success" size="text-5xl" />
